상점, 파티클추가
