+++ Assets/Game/Animation/Attack_AirSlash2.anim
@@ -125614,6 +125614,13 @@ AnimationClip:
     floatParameter: 0
     intParameter: 0
     messageOptions: 0
+  - time: 0.24
+    functionName: ParticleInstantiate
+    data: 
+    objectReferenceParameter: {fileID: 4118222606303123960, guid: 05ca08d7e09ce9b43837c3e6e4d5a513, type: 3}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
   - time: 1.12
     functionName: Attack_AirSlashEnd
     data:
