+++ Assets/Game/Prefabs/AttackParticle/Attack_AirSlash1.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 19cd1ffbce04138429b1d65929045d5b
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
