+++ Assets/Game/Prefabs/AttackParticle/Attack_AirSlash2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 05ca08d7e09ce9b43837c3e6e4d5a513
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
