+++ Assets/Game/Prefabs/AttackParticle/Attack_AirSlash3.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 24a2e370cd7341c4d8dea6b3356ab706
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
