+++ Assets/Game/Prefabs/HpPotionParticle.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: eef68f5e4e1e8c84ca19d70554f62c17
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
