+++ Assets/Game/Scripts/Item/EquipmentCombinationSystem.cs
@@ -33,24 +33,25 @@ public class EquipmentCombinationSystem : MonoBehaviour
     }
     private void Start()
     {
-        for (int i = 0; i < ItemListManager.instance.eqiupmentItemRecipeList.Count; i++)
+        for (int i = 0; i < JsonManager.instance.eqiupmentItemRecipeList.Count; i++)
         {
             CombiContentsUI temp = Instantiate(contentsPrifab, contents.transform).GetComponent<CombiContentsUI>();
-            temp.image.sprite = ItemListManager.instance.equipmentDic[ItemListManager.instance.eqiupmentItemRecipeList[i].result].sprite;
-            temp.textMeshProUGUI.text = ItemListManager.instance.eqiupmentItemRecipeList[i].result;
+            temp.image.sprite = JsonManager.instance.equipmentDic[JsonManager.instance.eqiupmentItemRecipeList[i].result].sprite;
+            temp.textMeshProUGUI.text = JsonManager.instance.eqiupmentItemRecipeList[i].result;
             int tempint = i;
             temp.gameObject.GetComponent<Button>().onClick.AddListener(() => { CombiRecipeView(tempint); });
             contentsRectTransform.sizeDelta = new Vector2(contentsRectTransform.sizeDelta.x, contentsRectTransform.sizeDelta.y + 50);
         }
+        playerGold.text = "보유 금액 : " + PlayerStatus.gold.ToString();
     }
 
     public void CombiRecipeView(int num)
     {
         target = num;
-        playerGold.text = PlayerStatus.gold.ToString();
-        EquipmentItem equipmentItem = ItemListManager.instance.equipmentDic[ItemListManager.instance.eqiupmentItemRecipeList[num].result];
-        MaterialItem materialItemA = ItemListManager.instance.materialsDic[ItemListManager.instance.eqiupmentItemRecipeList[num].materialA];
-        MaterialItem materialItemB = ItemListManager.instance.materialsDic[ItemListManager.instance.eqiupmentItemRecipeList[num].materialB];
+        playerGold.text ="보유 금액 : " + PlayerStatus.gold.ToString();
+        EquipmentItem equipmentItem = JsonManager.instance.equipmentDic[JsonManager.instance.eqiupmentItemRecipeList[num].result];
+        MaterialItem materialItemA = JsonManager.instance.materialsDic[JsonManager.instance.eqiupmentItemRecipeList[num].materialA];
+        MaterialItem materialItemB = JsonManager.instance.materialsDic[JsonManager.instance.eqiupmentItemRecipeList[num].materialB];
         int materialAnum = WarehouseManager.instance.FindItem(materialItemA);
         int materialBnum = WarehouseManager.instance.FindItem(materialItemB);
 
@@ -59,10 +60,10 @@ public class EquipmentCombinationSystem : MonoBehaviour
         materialBImage.sprite = materialItemB.sprite;
 
         resultName.text = equipmentItem.itemName;
-        materialAName.text = materialItemA.itemName + ItemListManager.instance.eqiupmentItemRecipeList[num].numA + " (" + materialAnum + ")";
-        materialBName.text = materialItemB.itemName + ItemListManager.instance.eqiupmentItemRecipeList[num].numB + " (" + materialBnum + ")";
-        needGold.text = ItemListManager.instance.eqiupmentItemRecipeList[num].gold.ToString();
-        if (materialAnum >= ItemListManager.instance.eqiupmentItemRecipeList[num].numA && materialBnum >= ItemListManager.instance.eqiupmentItemRecipeList[num].numB && PlayerStatus.gold >= ItemListManager.instance.eqiupmentItemRecipeList[num].gold)
+        materialAName.text = materialItemA.itemName + JsonManager.instance.eqiupmentItemRecipeList[num].numA + " (" + materialAnum + ")";
+        materialBName.text = materialItemB.itemName + JsonManager.instance.eqiupmentItemRecipeList[num].numB + " (" + materialBnum + ")";
+        needGold.text = JsonManager.instance.eqiupmentItemRecipeList[num].gold.ToString();
+        if (materialAnum >= JsonManager.instance.eqiupmentItemRecipeList[num].numA && materialBnum >= JsonManager.instance.eqiupmentItemRecipeList[num].numB && PlayerStatus.gold >= JsonManager.instance.eqiupmentItemRecipeList[num].gold)
             combinationButton.SetActive(true);
         else
             combinationButton.SetActive(false);
@@ -70,10 +71,10 @@ public class EquipmentCombinationSystem : MonoBehaviour
 
     public void Combination()
     {
-        WarehouseManager.instance.AddItem(ItemListManager.instance.equipmentDic[ItemListManager.instance.eqiupmentItemRecipeList[target].result]);
-        WarehouseManager.instance.MinusItem(ItemListManager.instance.materialsDic[ItemListManager.instance.eqiupmentItemRecipeList[target].materialA], ItemListManager.instance.eqiupmentItemRecipeList[target].numA);
-        WarehouseManager.instance.MinusItem(ItemListManager.instance.materialsDic[ItemListManager.instance.eqiupmentItemRecipeList[target].materialB], ItemListManager.instance.eqiupmentItemRecipeList[target].numB);
-        PlayerStatus.gold -= ItemListManager.instance.eqiupmentItemRecipeList[target].gold;
+        WarehouseManager.instance.AddItem  (JsonManager.instance.equipmentDic[JsonManager.instance.eqiupmentItemRecipeList[target].result   ]                                                           );
+        WarehouseManager.instance.MinusItem(JsonManager.instance.materialsDic[JsonManager.instance.eqiupmentItemRecipeList[target].materialA], JsonManager.instance.eqiupmentItemRecipeList[target].numA);
+        WarehouseManager.instance.MinusItem(JsonManager.instance.materialsDic[JsonManager.instance.eqiupmentItemRecipeList[target].materialB], JsonManager.instance.eqiupmentItemRecipeList[target].numB);
+        PlayerStatus.gold -= JsonManager.instance.eqiupmentItemRecipeList[target].gold;
         CombiRecipeView(target);
     }
 }
