+++ Assets/Game/Scripts/Item/Shop.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c1f0d020262178f4db5d3c5a23c1d896
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
