--- Assets/Game/Scripts/Manager/ItemListManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 900f91c141210f9448f5d911055d975b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
