+++ Assets/Game/Scripts/Manager/JsonManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d625eaee747ee154b86d04567e8397a4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
